Update Darts.java
This prevent rejection from the tests by default.

+++ exercises/practice/darts/src/main/java/Darts.java
@@ -1,5 +1,5 @@
 class Darts {
-    int score(int xOfDart, int yOfDart) {
+    int score(double xOfDart, double yOfDart) {
         throw new UnsupportedOperationException("Delete this statement and write your own implementation.");
     }
 }
